+++ contract/utxo.go
@@ -38,7 +38,7 @@ func PrevUtxoGetFunc(address unsafe.Pointer, context unsafe.Pointer) {
 	}
 
 	utxoLength := C.int(len(engine.prevUtxos))
-	utxosAddr := (*C.struct_utxo_t)(C.malloc(C.ulong(C.sizeof_struct_utxo_t * utxoLength)))
+	utxosAddr := (*C.struct_utxo_t)(C.malloc(C.size_t(C.sizeof_struct_utxo_t * utxoLength)))
 	defer C.free(unsafe.Pointer(utxosAddr))
 	utxos := (*[1 << 30]C.struct_utxo_t)(unsafe.Pointer(utxosAddr))[:utxoLength:utxoLength]
 	for index, prevUtxo := range engine.prevUtxos {
